backend/requirements.txt
Flask==2.3.7
flask-cors==4.0.0
flask-jwt-extended==4.4.4
flask-limiter==3.5.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
